+++ applications/minotari_node/src/lib.rs
@@ -35,7 +35,6 @@ mod grpc;
 mod grpc_method;
 #[cfg(feature = "metrics")]
 mod metrics;
-mod monitor_peers;
 mod recovery;
 mod utils;
 
@@ -169,13 +168,13 @@ pub async fn run_base_node_with_cli(
             "Force Sync Peers have been set! This node will only sync to the nodes in this set."
         );
     }
-    // Monitor peers to ensure the reported active connections are still active
-    task::spawn(monitor_peers::monitor_peers(
-        ctx.base_node_comms().clone(),
-        ctx.liveness(),
-        shutdown,
-        config.base_node.metadata_auto_ping_interval,
-    ));
+    // // Monitor peers to ensure the reported active connections are still active
+    // task::spawn(monitor_peers::monitor_peers(
+    //     ctx.base_node_comms().clone(),
+    //     ctx.liveness(),
+    //     shutdown,
+    //     config.base_node.metadata_auto_ping_interval,
+    // ));
 
     info!(target: LOG_TARGET, "Minotari base node has STARTED");
     main_loop.cli_loop().await;
