+++ base_layer/p2p/src/services/liveness/service.rs
@@ -20,7 +20,11 @@
 // WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
 // USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-use std::{iter, sync::Arc, time::Instant};
+use std::{
+    iter,
+    sync::Arc,
+    time::{Duration, Instant},
+};
 
 use futures::{future::Either, pin_mut, stream::StreamExt, Stream};
 use log::*;
@@ -54,6 +58,8 @@ use crate::{
     services::liveness::{handle::LivenessEventSender, LivenessEvent, PingPongEvent},
     tari_message::TariMessageType,
 };
+
+pub const MAX_INFLIGHT_TTL: Duration = Duration::from_secs(30);
 
 /// Service responsible for testing Liveness of Peers.
 pub struct LivenessService<THandleStream, TPingStream> {
@@ -131,9 +137,7 @@ where
                         warn!(target: LOG_TARGET, "Error when pinging peers: {}", err);
                     }
                     if self.config.max_allowed_ping_failures > 0 {
-                        if let Err(err) = self.disconnect_failed_peers().await {
-                            error!(target: LOG_TARGET, "Error occurred while disconnecting failed peers: {}", err);
-                        }
+                        self.disconnect_failed_peers().await;
                     }
                 },
 
@@ -240,7 +244,11 @@ where
     async fn send_ping(&mut self, node_id: NodeId) -> Result<u64, LivenessError> {
         let msg = PingPongMessage::ping_with_metadata(self.state.metadata().clone());
         let nonce = msg.nonce;
-        self.state.add_inflight_ping(nonce, node_id.clone());
+        self.state.add_inflight_ping(
+            nonce,
+            node_id.clone(),
+            self.config.auto_ping_interval.unwrap_or(MAX_INFLIGHT_TTL),
+        );
         debug!(target: LOG_TARGET, "Sending ping to peer '{}'", node_id.short_str(),);
 
         self.outbound_messaging
@@ -277,12 +285,11 @@ where
                 self.state.inc_pings_sent();
                 Ok(LivenessResponse::Ok(Some(vec![nonce])))
             },
-            SendPings(node_ids, delay_between_pings) => {
+            SendPings(node_ids) => {
                 let mut nonces = Vec::with_capacity(node_ids.len());
                 for node_id in node_ids {
                     nonces.push(self.send_ping(node_id).await?);
                     self.state.inc_pings_sent();
-                    time::sleep(delay_between_pings).await;
                 }
                 Ok(LivenessResponse::Ok(Some(nonces)))
             },
@@ -348,7 +355,11 @@ where
 
         for peer in selected_peers {
             let msg = PingPongMessage::ping_with_metadata(self.state.metadata().clone());
-            self.state.add_inflight_ping(msg.nonce, peer.clone());
+            self.state.add_inflight_ping(
+                msg.nonce,
+                peer.clone(),
+                self.config.auto_ping_interval.unwrap_or(MAX_INFLIGHT_TTL),
+            );
             self.outbound_messaging
                 .send_direct_node_id(
                     peer,
@@ -363,24 +374,31 @@ where
         Ok(())
     }
 
-    async fn disconnect_failed_peers(&mut self) -> Result<(), LivenessError> {
+    async fn disconnect_failed_peers(&mut self) {
         let max_allowed_ping_failures = self.config.max_allowed_ping_failures;
+        let mut node_ids = Vec::new();
         for node_id in self
             .state
             .failed_pings_iter()
             .filter(|(_, n)| **n > max_allowed_ping_failures)
             .map(|(node_id, _)| node_id)
         {
-            if let Some(mut conn) = self.connectivity.get_connection(node_id.clone()).await? {
+            if let Ok(Some(mut conn)) = self.connectivity.get_connection(node_id.clone()).await {
                 debug!(
                     target: LOG_TARGET,
                     "Disconnecting peer {} that failed {} rounds of pings", node_id, max_allowed_ping_failures
                 );
-                conn.disconnect(Minimized::No).await?;
+                match conn.disconnect(Minimized::No).await {
+                    Ok(_) => {
+                        node_ids.push(node_id.clone());
+                    },
+                    Err(err) => {
+                        warn!(target: LOG_TARGET, "Failed to disconnect peer {} ({})", node_id, err);
+                    },
+                }
             }
         }
-        self.state.clear_failed_pings();
-        Ok(())
+        self.state.clear_failed_pings(&node_ids);
     }
 
     fn publish_event(&mut self, event: LivenessEvent) {
@@ -628,6 +646,7 @@ mod test {
         state.add_inflight_ping(
             msg.inner.as_ref().map(|i| i.nonce).unwrap(),
             msg.source_peer.node_id.clone(),
+            MAX_INFLIGHT_TTL,
         );
 
         // A stream which emits an inflight pong message and an unexpected one
